requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pyaml==23.7.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.4
